refactor: address Copilot follow-up review findings
- Strip task/review provenance from comments (PR numbers, Codex P2 /
  CodeRabbit / Copilot tags) introduced in earlier commits on this
  branch. AGENTS.md asks comments to explain the "why" only; the
  context belongs in the PR description, not the source.

- Extract `nowLocalIsoTimestamp()` into `@pu-stats/models` so the
  dashboard component and the QuickAddOrchestrationService stop
  duplicating the same five-line helper (browser-local minute clock +
  appendLocalOffset). Includes unit tests for shape + padding.

- Remove the one-shot `tools/translate-pr360-units.mjs` script that
  was committed alongside the locale translations. Its header marks
  itself as single-use; the audit trail lives in commit history.
