+++ libs/stats/src/index.ts
@@ -6,6 +6,7 @@ export * from './lib/date/parse-iso-date';
 export * from './lib/date/to-local-iso-date';
 export * from './lib/date/to-berlin-iso-date';
 export * from './lib/date/append-local-offset';
+export * from './lib/date/now-local-iso-timestamp';
 export * from './lib/date/create-week-range';
 export * from './lib/date/infer-range-mode';
 export * from './lib/date/range-modes.type';
